# backend/docs/openapi-schema.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Multi-Agent Behavioral Testing Platform",
    "description": "\n    Enterprise-grade framework for evaluating multi-agent systems through behavioral test graphs.\n    \n    ## Features\n    - **Graph Management**: Upload and version YAML-defined behavioral test graphs\n    - **Test Execution**: Run tests with deterministic seed control and replay capability\n    - **Contract Validation**: Enforce contracts between agent nodes\n    - **Multi-Provider**: Support for OpenAI, Anthropic, Azure, Ollama, and more\n    - **Governance**: PII detection, policy enforcement, safety scoring\n    - **Metrics**: Latency tracking, cost accounting, drift detection\n    - **Async Execution**: Background workers with webhook notifications\n    ",
    "version": "1.0.0"
  },
  "paths": {
...
    "/runs/{run_id}/replay": {
      "post": {
        "tags": [
          "Test Runs"
        ],
        "summary": "Replay Run",
        "description": "Replay a historical run deterministically using stored graph version.",
        "operationId": "replay_run_runs__run_id__replay_post",
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Run Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/ExecutionConfig"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Overrides"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/user-testing/runs/history": {
      "get": {
        "tags": [
          "User Testing"
        ],
        "summary": "Run History",
        "description": "Return recent runs for the QA review interface.",
        "operationId": "run_history_user_testing_runs_history_get",
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/user-testing/runs/{run_id}/timeline": {
      "get": {
        "tags": [
          "User Testing"
        ],
        "summary": "Run Timeline",
        "description": "Produce an execution timeline for a run.",
        "operationId": "run_timeline_user_testing_runs__run_id__timeline_get",
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/user-testing/runs/{run_id}/assertions": {
      "get": {
        "tags": [
          "User Testing"
        ],
        "summary": "Run Assertions",
        "description": "Return assertion outcomes for a run.",
        "operationId": "run_assertions_user_testing_runs__run_id__assertions_get",
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/user-testing/runs/{run_id}/compliance": {
      "get": {
        "tags": [
          "User Testing"
        ],
        "summary": "Run Compliance",
        "description": "Return safety/compliance violations for a run.",
        "operationId": "run_compliance_user_testing_runs__run_id__compliance_get",
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/user-testing/runs/{run_id}/control": {
      "post": {
        "tags": [
          "User Testing"
        ],
        "summary": "Run Control",
        "description": "Apply a control action (pause/resume/stop/replay) to a run.",
        "operationId": "run_control_user_testing_runs__run_id__control_post",
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Run Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunControlRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/user-testing/simulations": {
      "post": {
        "tags": [
          "User Testing"
        ],
        "summary": "Simulate Run",
        "description": "Execute a graph simulation without persisting results.",
        "operationId": "simulate_run_user_testing_simulations_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SimulationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/user-testing/runs/{run_id}/replay": {
      "get": {
        "tags": [
          "User Testing"
        ],
        "summary": "Replay Trace",
        "description": "Return the persisted execution trace for replay.",
        "operationId": "replay_trace_user_testing_runs__run_id__replay_get",
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ExecutionConfig": {
        "properties": {
          "provider": {
            "type": "string",
            "title": "Provider",
            "default": "mock"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "seed": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seed"
          },
          "mode": {
            "type": "string",
            "title": "Mode",
            "default": "normal"
          },
          "webhook_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Webhook Url"
          },
          "chaos_config": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chaos Config"
          },
          "timeout_seconds": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timeout Seconds"
          },
          "tenant_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tenant Id"
          },
          "governance": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Governance"
          }
        },
        "type": "object",
        "title": "ExecutionConfig",
        "description": "Configuration for test execution."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RunControlAction": {
        "type": "string",
        "enum": [
          "pause",
          "resume",
          "stop",
          "replay"
        ],
        "title": "RunControlAction",
        "description": "Supported user testing actions on a run."
      },
      "RunControlRequest": {
        "properties": {
          "action": {
            "$ref": "#/components/schemas/RunControlAction"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          }
        },
        "type": "object",
        "required": [
          "action"
        ],
        "title": "RunControlRequest"
      },
      "SimulationRequest": {
        "properties": {
          "graph": {
            "additionalProperties": true,
            "type": "object",
            "title": "Graph"
          },
          "execution_config": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Execution Config"
          }
        },
        "type": "object",
        "required": [
          "graph"
        ],
        "title": "SimulationRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      },
      "APIKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  }
}